refactor move lock widget and JSON related classes inside an internal pkg
