+++ lib/src/main/java/com/auth0/android/lock/Lock.java
@@ -41,6 +41,9 @@ import com.auth0.android.lock.LockCallback.LockEvent;
 import com.auth0.android.lock.enums.InitialScreen;
 import com.auth0.android.lock.enums.SocialButtonStyle;
 import com.auth0.android.lock.enums.UsernameStyle;
+import com.auth0.android.lock.utils.CustomField;
+import com.auth0.android.lock.internal.Options;
+import com.auth0.android.lock.internal.Theme;
 import com.auth0.android.lock.provider.AuthProviderResolver;
 import com.auth0.android.lock.provider.ProviderResolverManager;
 import com.auth0.android.lock.utils.LockException;
