+++ lib/src/main/java/com/auth0/android/lock/LockActivity.java
@@ -56,9 +56,11 @@ import com.auth0.android.lock.events.DatabaseSignUpEvent;
 import com.auth0.android.lock.events.EnterpriseLoginEvent;
 import com.auth0.android.lock.events.FetchApplicationEvent;
 import com.auth0.android.lock.events.SocialConnectionEvent;
+import com.auth0.android.lock.internal.Configuration;
+import com.auth0.android.lock.internal.Options;
 import com.auth0.android.lock.provider.ProviderResolverManager;
-import com.auth0.android.lock.utils.json.ApplicationFetcher;
-import com.auth0.android.lock.utils.json.Connection;
+import com.auth0.android.lock.internal.json.ApplicationFetcher;
+import com.auth0.android.lock.internal.json.Connection;
 import com.auth0.android.lock.views.ClassicLockView;
 import com.auth0.android.provider.AuthCallback;
 import com.auth0.android.provider.AuthProvider;
