+++ lib/src/main/java/com/auth0/android/lock/PasswordlessLockActivity.java
@@ -57,9 +57,11 @@ import com.auth0.android.lock.events.CountryCodeChangeEvent;
 import com.auth0.android.lock.events.FetchApplicationEvent;
 import com.auth0.android.lock.events.PasswordlessLoginEvent;
 import com.auth0.android.lock.events.SocialConnectionEvent;
+import com.auth0.android.lock.internal.Configuration;
+import com.auth0.android.lock.internal.Options;
 import com.auth0.android.lock.provider.ProviderResolverManager;
-import com.auth0.android.lock.utils.json.ApplicationFetcher;
-import com.auth0.android.lock.utils.json.Connection;
+import com.auth0.android.lock.internal.json.ApplicationFetcher;
+import com.auth0.android.lock.internal.json.Connection;
 import com.auth0.android.lock.views.PasswordlessLockView;
 import com.auth0.android.provider.AuthCallback;
 import com.auth0.android.provider.AuthProvider;
