+++ lib/src/main/java/com/auth0/android/lock/internal/json/ApplicationDeserializer.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.internal.json;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
