+++ lib/src/main/java/com/auth0/android/lock/internal/json/ApplicationFetcher.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.internal.json;
 
 import android.net.Uri;
 import android.support.annotation.NonNull;
