+++ lib/src/main/java/com/auth0/android/lock/internal/json/Strategy.java
@@ -1,4 +1,4 @@
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.internal.json;
 
 import android.support.annotation.NonNull;
