+++ lib/src/main/java/com/auth0/android/lock/utils/EnterpriseConnectionMatcher.java
@@ -28,8 +28,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
-import com.auth0.android.lock.enums.AuthType;
-import com.auth0.android.lock.utils.json.Connection;
+import com.auth0.android.lock.internal.json.Connection;
 
 import java.util.ArrayList;
 import java.util.List;
