+++ lib/src/main/java/com/auth0/android/lock/utils/LoadCountriesTask.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.utils;
 
 import android.content.Context;
 import android.os.AsyncTask;
