+++ lib/src/main/java/com/auth0/android/lock/views/ActionButton.java
@@ -38,7 +38,7 @@ import android.widget.ImageView;
 import android.widget.ProgressBar;
 
 import com.auth0.android.lock.R;
-import com.auth0.android.lock.Theme;
+import com.auth0.android.lock.internal.Theme;
 
 public class ActionButton extends FrameLayout {
