+++ lib/src/main/java/com/auth0/android/lock/views/CountryCodeSelectorView.java
@@ -38,7 +38,7 @@ import android.widget.TextView;
 
 import com.auth0.android.lock.R;
 import com.auth0.android.lock.adapters.Country;
-import com.auth0.android.lock.utils.json.LoadCountriesTask;
+import com.auth0.android.lock.utils.LoadCountriesTask;
 
 import java.util.ArrayList;
 import java.util.Locale;
