+++ lib/src/main/java/com/auth0/android/lock/views/ValidatedUsernameInputView.java
@@ -27,7 +27,7 @@ package com.auth0.android.lock.views;
 import android.content.Context;
 import android.util.AttributeSet;
 
-import com.auth0.android.lock.Configuration;
+import com.auth0.android.lock.internal.Configuration;
 
 import static com.auth0.android.lock.enums.UsernameStyle.DEFAULT;
 import static com.auth0.android.lock.enums.UsernameStyle.EMAIL;
