+++ lib/src/test/java/com/auth0/android/lock/internal/json/ApplicationFetcherTest.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.internal.json;
 
 import com.auth0.android.Auth0;
 import com.auth0.android.Auth0Exception;
