+++ lib/src/test/java/com/auth0/android/lock/internal/json/ConnectionMatcher.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.utils.json;
+package com.auth0.android.lock.internal.json;
 
 
 import org.hamcrest.BaseMatcher;
